Split out api
